% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{determine_marginals}
\alias{determine_marginals}
\title{Title}
\usage{
determine_marginals(bayesWatch_object, model_logs, prob_cutoff)
}
\arguments{
\item{variable_names}{}
}
\description{
Title
}
