% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regimes.R
\name{fit_regime_vector}
\alias{fit_regime_vector}
\title{Title}
\usage{
fit_regime_vector(
  data_woTimeValues,
  time_of_observations,
  time_points,
  variable_names = 1:ncol(data_woTimeValues),
  not.cont = NULL,
  iterations = 100,
  burnin = floor(iterations/2),
  lower_bounds = NULL,
  upper_bounds = NULL,
  previous_states = NULL,
  previous_model_fits = NULL,
  sparsity_parameter = NULL,
  linger_parameter = 10,
  move_parameter = 1,
  g.prior = 0.2,
  set_G = NULL,
  wishart_df_inital = 20,
  lambda = 20,
  g_sampling_distribution = NULL,
  n.cores = 1,
  scaleMatrix = NULL,
  allow_for_mixture_models = FALSE,
  dirichlet_prior = 0.004,
  component_truncation = 7,
  regime_truncation = 15,
  hyperprior_b = 20,
  model_params_save_every = 5,
  simulation_iter = NULL,
  T2_window_size = 3,
  determining_p_cutoff = FALSE,
  prob_cutoff = 0.5,
  verbose_logfile = TRUE
)
}
\arguments{
\item{determining_p_cutoff}{}
}
\description{
Title
}
