% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{full_data}
\alias{full_data}
\alias{day_of_observations}
\alias{day_dts}
\title{Simulated Data with Imposed Change-points.}
\format{
`full_data` is a matrix, the latter two are vectors.
}
\description{
Data simulated using the BDgraph package.  A change-point is imposed between days 5
and 6.  The change only occurs in variables 3 and 4.
}
\details{
`full_data` is a data frame with 1,000 rows and 5 columns.
`day_of_observations`; is a timestamp of each of `full_data`'s 1,000 rows.
`day_dts`; is a vector of unique elements from `day_of_observations`..
}
\examples{
data("full_data", package = "bayesWatch")
head(full_data)
data("day_of_observations", package = "bayesWatch")
head(day_of_observations)
data("day_dts", package = "bayesWatch")
head(day_dts)
}
\keyword{datasets}
