% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{graph_posterior_distributions}
\alias{graph_posterior_distributions}
\title{Title}
\usage{
graph_posterior_distributions(
  differences_of_marginals,
  bayesWatch_object,
  model_logs,
  prob_cutoff,
  f_divergence = "Hellinger"
)
}
\arguments{
\item{simulation_number}{}
}
\description{
Title
}
