% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regimes.R
\name{plot.bayesWatch}
\alias{plot.bayesWatch}
\title{Plots posterior probabilities of a change-point.}
\usage{
\method{plot}{bayesWatch}(x, ...)
}
\arguments{
\item{x}{bayesWatch object. Fit with the bayesWatch method.}

\item{...}{Additional plotting arguments.}
}
\value{
a ggplot.
}
\description{
Plots posterior probabilities of a change-point.
}
