% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regimes.R
\name{plot.bayesWatch}
\alias{plot.bayesWatch}
\title{Print function for a bayesWatch object.  Prints only the posterior change-point probabilities.}
\arguments{
\item{x}{bayesWatch object. Fit from bayesWatch main method.}

\item{...}{Additional plotting arguments.}
}
\description{
Print function for a bayesWatch object.  Prints only the posterior change-point probabilities.
}
